+++ test/test_parameters.py
@@ -1,5 +1,6 @@
 import datetime
 from copy import deepcopy
+from urllib.parse import quote, unquote
 
 import pytest
 import yaml